+++ Engine/Scene/Task/TaskComponentUpdateCamera.h
@@ -31,11 +31,11 @@ public:
 	virtual void TaskFunc(void *pParams)
 	{
 		if (m_pComponentManager) {
-			uint32_t count = (m_pComponentManager->GetComponentCount() + (THREAD_COUNT - m_pComponentManager->GetComponentCount() % THREAD_COUNT)) / THREAD_COUNT;
-			uint32_t indexBegin = max(count * m_indexThread, 0);
-			uint32_t indexEnd = min(count * (m_indexThread + 1), m_pComponentManager->GetComponentCount());
+			size_t count = (m_pComponentManager->GetComponentCount() + (THREAD_COUNT - m_pComponentManager->GetComponentCount() % THREAD_COUNT)) / THREAD_COUNT;
+			size_t indexBegin = max(count * m_indexThread, 0);
+			size_t indexEnd = min(count * (m_indexThread + 1), m_pComponentManager->GetComponentCount());
 
-			for (uint32_t index = indexBegin; index < indexEnd; index++) {
+			for (size_t index = indexBegin; index < indexEnd; index++) {
 				CComponent *pComponent = (CComponent *)m_pComponentManager->GetComponentByIndex(index);
 				pComponent->UpdateCamera(m_pCamera, m_indexThread);
 			}
